Split out cynic-proc-macros & fix scalars_as_strings.
