+++ cynic/Cargo.toml
@@ -15,7 +15,7 @@ documentation = "https://docs.rs/cynic"
 json-decode = "0.2.0"
 serde = { version = "1.0.104", features = [ "derive" ] }
 serde_json = "1.0.44"
-cynic-codegen = { path = "../cynic-codegen", version = "0.1.2" }
+cynic-proc-macros = { path = "../cynic-proc-macros", version = "0.1.2" }
 
 [dev-dependencies]
 maplit = "1.0.2"
